% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basals.R
\name{basals}
\alias{basals}
\title{Baseline Levels}
\usage{
basals(df, patientCol, visitCol, visit0, bioCols)
}
\arguments{
\item{df}{The dataframe with the expression levels of the biomarkers.}

\item{patientCol}{The column name with the information of the patients (e.g., "patients").}

\item{visitCol}{The column name with the information of the visits (e.g., "visits").}

\item{visit0}{The name of the first visit (which contains baseline info) (e.g., "Visit 0").}

\item{bioCols}{Range of columns that contain the biomarkers (e.g., 5:500).}
}
\value{
A dataframe with the original columns and the baseline expression columns.
}
\description{
This function returns a dataframe with the baseline levels of the selected
biomarkers, adding new columns with those baseline levels with the name of
the metabolite and the suffix "_basal". It has five arguments: the provided
dataframe, the name of the column in which the patient information is stored
(in order to group by this variable, as each patient has different baseline
levels for each biomarker), the column in which the visit information is saved,
the name of the first visit (e.g., "FPE1 ANTES") and, finally, the range of the
biomarkers columns (e.g., if biomarkers levels begin at column 5 and finish at
35, this parameter should be 5:35).
}
