% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMap.R
\name{corrMap}
\alias{corrMap}
\title{Correlation Maps}
\usage{
corrMap(df, sig = NULL, type = NULL)
}
\arguments{
\item{df}{A provided dataframe.}

\item{sig}{Significance threshold (e.g., 0.05). Null by default.}

\item{type}{Type of correlation. It accepts "pearson" or "spearman".
NULL by default ("pearson").}
}
\value{
A correlation plot.
}
\description{
This function generates a customized correlation plot from a provided
dataframe. It will extract the numeric variables from the dataframe and it
will return a plot, highlighting positive correlations in red and negative
correlations in blue. In addition, we can provide a desired significance
threshold and it will only color those correlations that are under the
targeted level of significance. It only needs two arguments: a dataframe and
a significance threshold (NULL by default).
}
