% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUCRows.R
\name{findUCRows}
\alias{findUCRows}
\title{Find Unique and Common Rows}
\usage{
findUCRows(dataframes, column)
}
\arguments{
\item{dataframes}{List of dataframes.}

\item{column}{Column in which we have to look the unique/common values.}
}
\value{
A list with the unique and common rows.
}
\description{
This is a helper function that is designed to find those rows that are
unique for a given dataframe or that are shared among a list of dataframes.
This function looks at a provided column and if the values for that column
are unique in a certain dataframe, it stores those rows as unique for that
dataframe; on the other hand, if it finds common values among the provided
dataframes, it stores them in another list. Thus, it only needs two
arguments: the list of dataframes and the column in which it has to look
those values. It returns a list with the unique and common rows that can
be used by other functions.
}
