% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metBarPlot.R
\name{metBarPlot}
\alias{metBarPlot}
\title{Bar Plot of Metabolite Class by Significance}
\usage{
metBarPlot(toptable, class, estimate, adj, sigs, title = NULL, subt = NULL)
}
\arguments{
\item{toptable}{The top table with the necessary information.}

\item{class}{Name of the 'class' column in the top table.}

\item{estimate}{Name of the 'estimate' column in the top table.}

\item{adj}{Name of the 'adjusted p-value' column in the top table.}

\item{sigs}{Vector with the desired adjusted p-values thresholds.}

\item{title}{Title of the bar plot (NULL by default)}

\item{subt}{Subtitle of the bar plot (NULL by default)}
}
\value{
A bar plot with the percentage of biomarkers up-regulated or down-
regulated depending on the significance level.
}
\description{
This functions is a variation of metTables. Instead of a table with
the different N and percentatge of the biomarker Classes, it returns
a bar plot with that information: the percentage at the Y axis and the
count (n) at a given significance level specified inside the bar.
It has five arguments: the toptable, the names of the 'class', 'estimate'
and 'adjusted p-value' columns in that toptable and the significance
thresholds desired by the user. It returns a bar plot with the percentage of
biomarkers up-regulated or down-regulated depending on the significance level.
(if not, it returns a brief message communicating the lack of biomarkers at
those levels).
}
