% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metTables.R
\name{metTables}
\alias{metTables}
\title{Table of Metabolite Class by Significance}
\usage{
metTables(toptable, class, estimate, adj, sigs)
}
\arguments{
\item{toptable}{The top table with the necessary information.}

\item{class}{Name of the 'class' column in the top table.}

\item{estimate}{Name of the 'estimate' column in the top table.}

\item{adj}{Name of the 'adjusted p-value' column in the top table.}

\item{sigs}{Vector with the desired adjusted p-values thresholds.}
}
\value{
A kable-formatted table with the number of differentially expressed biomarkers
by sig. level and class.
}
\description{
This functions returns a table with the different classes of biomarkers
in a given top-table and the number of total counts and the counts for
each significance level (provided by the user). It also calculates the
percentages of counts for each significance threshold with respect to the
total count for that class and divides the biomarkers in upregulated
and downregulated. Thus, it provides a general view of which classes
are being upregulated or downregulated. It has five arguments: the toptable,
the names of the 'class', 'estimate' and 'adjusted p-value' columns in that
toptable and the significance thresholds desired by the user. It returns a
table with all the information if there are biomarkers upregulated or down-
regulated below the provided significance thresholds (if not, it returns
a brief message communicating the lack of biomarkers at those levels).
}
