% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(df, label = FALSE)
}
\arguments{
\item{df}{A given dataframe}

\item{label}{Individual label for each outlier (by default it is FALSE).}
}
\value{
A dataframe with all the outliers found.
}
\description{
This function is designed to find all the outliers that are present in
a given dataframe. It iterates through every numeric column to get these
outliers values, which can be described as data points that differ
significantly from other observations. The function defines an outlier
as any value that is 1.5 times the interquantile range (IQR)  below the lower
quartile or 1.5IQR above the upper quartile. The function generates a boxplot
for each variable and highlights the outliers in red. In addition, the user
has the option of labelling the outliers by setting the argument label as
TRUE (it is FALSE by default), which individually labels each outlier with an
unique name and it also adds a subtitle to the plot with the outliers' values.
Finally, the function stores all the outliers detected in the dataframe in a
new dataframe, which it is also the object that it returns.
}
