% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigTables.R
\name{sigTables}
\alias{sigTables}
\title{Table of Significant Biomarkers}
\usage{
sigTables(toptable, sigs, estimate, pval, adj)
}
\arguments{
\item{toptable}{The top table with the necessary information.}

\item{sigs}{Vector with the desired significance levels (e.g., c(0.01, 0.05)).}

\item{estimate}{Name of the 'estimate' column in the top table.}

\item{pval}{Name of the 'raw p-value' column in the top table.}

\item{adj}{Name of the 'adjusted p-value' column in the top table.}

\item{visit0}{The name of the first visit (which contains baseline info) (e.g., "Visit 0").}

\item{bioCols}{Range of columns that contain the biomarkers (e.g., 5:500).}
}
\value{
A kable-formatted table with the number of significant biomarkers by sig. level.
}
\description{
This function returns a table with the number of significant biomarkers of a
provided top table. It accepts as many levels of significance as desired
by the user. In addition, it can work with both adjusted and raw p-values
as long as the user provides the name of their columns in the top table.
It has five arguments: the name of the toptable, the desired significance
levels and the names of the estimate, raw p-value and adjusted p-value columns.
It returns a kable-formatted table with the desired information.
}
