% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plots}
\usage{
volcanoPlot(endtable, estimate, adj, biomarker, estimateVal, adjVal)
}
\arguments{
\item{endtable}{Top-Table with the biomarker's data.}

\item{estimate}{Name of the "estimate" column in the top-table.}

\item{adj}{Name of the "adjusted p-value" column in the top-table.}

\item{biomarker}{Name of the "biomarker" column in the top-table.}

\item{estimateVal}{Absolute estimate value threshold for up or down-regulation.}

\item{adjVal}{Adjusted p-value threshold.}
}
\value{
An interactive ggplotly object with the Volcano plot.
}
\description{
This function will generate a Volcano plot from a provided top table. It will
highlight the differentially expressed biomarkers that are down or up-regulated.
The user can provide the estimate threshold from which a biomarker will be
up or down-regulated, as well as the desired adjusted p-value threshold.
The function has six arguments: the top-table with the data, the name
of the estimate column in that top-table, the name of the adjusted p-value
in the top-table, the name of the Biomarker column (i.e., the column that
contains the name or identification of each biomarker) and then the desired
thresholds for the estimate and adjusted p-value. It returns an interactive
ggplotly object that allows the user to focus on the relevant biomarkers, as the basic information of the highlighted
biomarkers will be displayed on mouseover.
}
