% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrmTables.R
\name{mmrmTable}
\alias{mmrmTable}
\title{MMRM Table}
\usage{
mmrmTable(model, df, bioCols, visits, excel = NULL)
}
\arguments{
\item{model}{MMRM model that we have generated before}

\item{df}{The dataframe that we used to create the model.}

\item{bioCols}{Range of columns that contain the biomarkers (e.g., 5:500).}

\item{excel}{Name of the Excel file which will contain the table data. Default NULL.}
}
\value{
A Kable-formatted table with the relevant information of the model.
}
\description{
This function will create a Kable-formatted table with the most relevant
biomarkers in a given model. It has four arguments: a model that has to be
provided; a dataframe in which we have the data that we used to run the model;
the range of columns in which we have the biomarker information; and an 'excel'
argument that expects the name of the desired file and which is null by default
(i.e., it will not create an Excel file with the information of the table
if we don't specify it).
}
