% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{vennDiagram}
\alias{vennDiagram}
\title{Venn diagrams}
\usage{
vennDiagram(
  endtable1,
  endtable2,
  estimate1,
  estimate2,
  adj1,
  adj2,
  biomarker,
  estimateVal,
  adjVal
)
}
\arguments{
\item{endtable1}{The first top-table.}

\item{endtable2}{The second top-table.}

\item{estimate1}{Name of the "estimate" column in the first top-table.}

\item{estimate2}{Name of the "estimate" column in the second top-table.}

\item{adj1}{Name of the "adjusted p-value" column in the first top-table.}

\item{adj2}{Name of the "adjusted p-value" column in the second top-table.}

\item{biomarker}{Name of the "biomarker" column in both top-tables.}

\item{estimateVal}{Absolute estimate value threshold for up or down-regulation.}

\item{adjVal}{Adjusted p-value threshold.}
}
\description{
This function creates a Venn Diagram to compare two top-tables. It also
generates a table with the shared biomarkers between the two top tables,
as well as two tables for the unique biomarkers of each top-tables. IT IS
ESSENTIAL TO RUN THE CHUNK WITH 'results = "asis"' IN ORDER TO VIEW THE
KABLE TABLES. It has nine arguments: the two top-tables that are going to
be compared, the name of the estimate, adjusted P-value and biomarker
columns and, finally, the desired estimate and adjusted P-value thresholds.
}
